Initial commit
This GitHub project showcases my implementation of stack and queue data structures in C# through the use of the MyStack and MyQueue generic classes. As a software engineer, I'm dedicated to refining my coding skills, and this project serves as a practical exploration of fundamental data structures. If you have any questions or if there's an aspect of the implementation you'd like to discuss, feel free to engage. I'm always open to learning and improving as an engineer!